+++ src/Regressors/KDNRegressor.php
@@ -4,10 +4,10 @@ namespace Rubix\ML\Regressors;
 
 use Rubix\ML\Learner;
 use Rubix\ML\Persistable;
+use Rubix\ML\Graph\KDTree;
 use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Datasets\Labeled;
 use Rubix\ML\Datasets\DataFrame;
-use Rubix\ML\Graph\KDTree;
 use Rubix\ML\Other\Helpers\Stats;
 use Rubix\ML\Kernels\Distance\Distance;
 use InvalidArgumentException;
@@ -40,27 +40,37 @@ class KDNRegressor extends KDTree implements Learner, Persistable
     protected $k;
 
     /**
+     * Should we use the inverse distances as confidence scores when
+     * making predictions?
+     * 
+     * @var bool
+     */
+    protected $weighted;
+
+    /**
      * @param  int  $k
-     * @param  int  $neighborhood
+     * @param  int  $maxLeafSize
      * @param  \Rubix\ML\Kernels\Distance\Distance|null  $kernel
+     * @param  bool  $weighted
      * @throws \InvalidArgumentException
      * @return void
      */
-    public function __construct(int $k = 3, int $neighborhood = 20, ?Distance $kernel = null)
+    public function __construct(int $k = 3, int $maxLeafSize = 20, ?Distance $kernel = null, bool $weighted = true)
     {
         if ($k < 1) {
-            throw new InvalidArgumentException("At least 1 neighbor is required"
+            throw new InvalidArgumentException('At least 1 neighbor is required'
                 . " to make a prediction, $k given.");
         }
 
-        if ($k > $neighborhood) {
-            throw new InvalidArgumentException("K cannot be larger than the max"
-                . " size of a neighborhood, $k given but $neighborhood allowed.");
+        if ($k > $maxLeafSize) {
+            throw new InvalidArgumentException('K cannot be larger than the max'
+                . " leaf size, $k given but $maxLeafSize allowed.");
         }
 
         $this->k = $k;
+        $this->weighted = $weighted;
 
-        parent::__construct($neighborhood, $kernel);
+        parent::__construct($maxLeafSize, $kernel);
     }
 
     /**
@@ -115,9 +125,21 @@ class KDNRegressor extends KDTree implements Learner, Persistable
         $predictions = [];
 
         foreach ($dataset as $sample) {
-            list($labels, $distances) = $this->neighbors($sample, $this->k);
+            list ($distances, $labels) = $this->neighbors($sample);
 
-            $predictions[] = Stats::mean($labels);
+            if ($this->weighted) {
+                $weights = [];
+
+                foreach ($distances as $i => $distance) {
+                    $weights[] = 1. / (1. + $distance);
+                }
+
+                $outcome = Stats::weightedMean($labels, $weights);
+            } else {
+                $outcome = Stats::mean($labels);
+            }
+
+            $predictions[] = $outcome;
         }
 
         return $predictions;